Finished todo and todo list reresources and tests

+++ app/todos/resources.py
@@ -15,38 +15,51 @@ class TodoMultiResource(Resource):
         response.
         """
         userId = session.get('userId')
+        permissions = TodoListPermission.query.filter_by(userId=userId)
         todoListId = request.args.get('todoListId')
 
+        if todoListId is None:
 
-        if filteredTodos is None:
+            todoListIds = [permission.todoListId for permission in permissions]
+            if not todoListIds:
+                return response_util.buildOkResponse([])
 
-            todoLists = [permission.todoList for permission in
-                         TodoListPermission.query.filter_by(userId=userId)]
-            myTodos = []
-            for todoList in todoLists:
-                myTodos.extend(todoList.todos)
+            myTodos = Todo.query.filter(Todo.todoListId.in_(todoListIds))
 
             return response_util.buildOkResponse([todo.toDict() for todo in myTodos])
         else:
             todoListId = int(todoListId)  # bad request if not int
             permissions = TodoListPermission.query.filter(
-                           and_(TodoListPermission.todoListId=todoListId,
-                                TodoListPermission.userId=userId))
+                           and_(TodoListPermission.todoListId == todoListId,
+                                TodoListPermission.userId == userId)).all()
 
-            if permissions == []:
-                raise status.Unauthorized() # change to new permission denied request
+            if not permissions:
+                raise status.Forbidden()
 
-            todos = Todos.query.filter_by(todoListId=todoListId)
+            todos = Todo.query.filter_by(todoListId=todoListId)
 
             return response_util.buildOkResponse([todo.toDict() for todo in todos])
 
+    @authentication.requiresAuth
     def post(self):
         """Method adds a new todo.
         """
+        userId = session.get('userId')
+        todoListId = request.form.get('todoListId')
+
         if not(request.form.get('subject') and request.form.get('todoListId')):
             raise status.BadRequest()
+
+        permission = TodoListPermission.query.filter(
+                           and_(TodoListPermission.todoListId == todoListId,
+                                TodoListPermission.userId == userId)).first()
+
+        if permission is None:
+            raise status.Forbidden()
+
         todo = Todo(request.form.get('subject'),
                     request.form.get('todoListId'),
+                    userId,
                     request.form.get('dueDate'),
                     request.form.get('description'),
                     request.form.get('priority'),
@@ -61,16 +74,25 @@ class TodoMultiResource(Resource):
 class TodoResource(Resource):
     """Class for updating, delelting, getting single todo.
     """
+    @authentication.requiresAuth
     def put(self, todoId):
         """Method updates a todo information and returns todo in an OK response.
 
         Args:
             todoId - Interger, primary key identifying the todo.
         """
+        userId = session.get('userId')
         todo = Todo.query.get(todoId)
 
         if todo is None:
             raise status.NotFound()
+
+        permission = TodoListPermission.query.filter(
+                           and_(TodoListPermission.todoListId == todo.todoListId,
+                                TodoListPermission.userId == userId)).first()
+
+        if permission is None:
+            raise status.Forbidden()
 
         todo.subject = request.form.get('subject') or todo.subject
         todo.dueDate = request.form.get('dueDate') or todo.dueDate
@@ -83,26 +105,48 @@ class TodoResource(Resource):
 
         return response_util.buildOkResponse(todo.toDict())
 
+    @authentication.requiresAuth
     def get(self, todoId):
         """Method gets and returns a single todo in an OK response.
 
         Args:
             todoId - Interger, primary key identifying the todo.
         """
+        userId = session.get('userId')
         todo = Todo.query.get(todoId)
+
         if todo is None:
             raise status.NotFound()
+
+        permission = TodoListPermission.query.filter(
+                           and_(TodoListPermission.todoListId == todo.todoListId,
+                                TodoListPermission.userId == userId)).first()
+
+        if permission is None:
+            raise status.Forbidden()
+
         return response_util.buildOkResponse(todo.toDict())
 
+    @authentication.requiresAuth
     def delete(self, todoId):
         """Method deletes a todo and returns result none.
 
         Args:
             todoId - Interger, primary key identifying the todo.
         """
+        userId = session.get('userId')
         todo = Todo.query.get(todoId)
+
         if todo is None:
             raise status.NotFound()
+
+        permission = TodoListPermission.query.filter(
+                           and_(TodoListPermission.todoListId == todo.todoListId,
+                                TodoListPermission.userId == userId)).first()
+
+        if permission is None:
+            raise status.Forbidden()
+
         db.session.delete(todo)
         db.session.commit()
